trying to prevent dead ppl from reading

+++ chatclient.c
@@ -15,19 +15,23 @@ int main() {
 int chat(int server) {
   printf("You have entered the chatroom!\n");
   char input[100];
-  int f = fork();
+  int f = 0;
 
-  if (f == 0) { // child waits for input to send
-    while (read(STDIN_FILENO, input, sizeof(input))) {
-      write(server, input, 100);
+  if (read(server,input,sizeof(input)) && !strcmp(input, "DEAD")){
+    f = fork();
+
+    if (f == 0) { // child waits for input to send
+      while (read(STDIN_FILENO, input, sizeof(input))) {
+        write(server, input, 100);
+      }
     }
   }
 
   // main program reads from server client msgs
-  while(read(server, input, sizeof(input))){
+  while(read(server, input, sizeof(input)) && strcmp(input, "STOPTALKING")){
     printf("%s", input);
   }
-  kill(f, SIGKILL); // removes child process
-  printf("chatroom over\n");
+  if (f) kill(f, SIGKILL); // removes child process
+  printf("\nchatroom over\n\n");
   return 0;
 }
